Move EvalContextGuard to dedicated eval-context.hh header
Separates the evaluation context mechanism from error.hh.
Call sites now include nix/util/eval-context.hh instead of
nix/util/error.hh, since the guard is about context, not errors.
error.hh still includes eval-context.hh for currentEvalContext()
used in BaseError constructors.
